Package checks to avoid import cycle
